Multi-Threading in C
Write a multi-threaded program that calculates various statistical values for a list of numbers. This program will be passed a series of numbers on the command line and will then create three separate worker threads. One thread will determine the  average of the numbers, the second will determine the maximum value, and the third will determine the minimum value. For example, suppose your program is passed the integers